Reorder index.html comments/code per requirements
The index2.html now utilizes functions for 2 reasons:
    1. Clarity
    2. Matching the Odin Project solution requirements

The Odin Project will ask us to revisit the codebase in the future for
a GUI solution.

Since index.html has a working solution but utilizes loops, this is
simply a reordering & tweaking already functioning code to use functions
but without needing loops.