Switch getPlayerChoice() to button listeners
The old getPlayerChoice() function is suitable for Rock Paper Scissors
that runs on the terminal but not for one with an UI.

The theorized reason behind this interaction is as follows.
If buttons w/ event listeners were inside of the getPlayerChoice()
function, the 1st round will not start from the 1st click because the
buttons' event listeners are added to the buttons once the first click
happen without making "humanChoice = buttonId";

The temporary solution is to have buttons w/ event listeners without
being inside a function. It fixes the issue and the code is semantically
self-explainatory along with the comments that follows.